# api/openapi.yaml
openapi: 3.0.3
info:
  title: Creatorr API
  version: 0.1.0
  description: |
    REST API for Creatorr (creator VOD daemon).
    HTMX HTML routes and SSE (`/api/events`) are out of this contract - see AGENTS.md.
servers:
  - url: /
    description: Local / Docker
...
paths:
...
  /api/series/{id}:
    get:
      tags: [library]
      operationId: getSeries
      summary: Series detail with sources and videos
      parameters:
        - $ref: "#/components/parameters/SeriesId"
      responses:
        "200":
          description: Series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Series"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [library]
      operationId: deleteSeries
      summary: Delete a series and its sources/indexed entries (library files on disk are kept; UI delete always removes library files after confirm)
      parameters:
        - $ref: "#/components/parameters/SeriesId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
...
    SeriesId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine code (e.g. CookieInvalid, DownloadFailed, RemuxFailed, PackFailed)
        message:
          type: string
          description: Short user-facing text
        detail:
          type: string
          description: Optional operator/debug detail
...
    Source:
      type: object
      required: [id, series_id, url, kind, monitored]
      properties:
        id:
          type: integer
          format: int64
        series_id:
          type: integer
          format: int64
        url:
          type: string
        label:
          type: string
          nullable: true
          description: Optional display name for the source (UI shows as Name)
        kind:
          type: string
          enum: [feed, single]
          description: feed = channel/playlist (catch-up); single = one video URL (index once, no catch-up)
        monitored:
          type: boolean
          description: Included in series scans when true (always true for kind=single; UI hides the toggle)
        title_regexp_include:
          type: string
          nullable: true
          description: >
            Optional Go regexp; new videos whose title does not match are not indexed.
            Empty = no include filter. Evaluated with title_regexp_exclude (exclude wins when both match).
        title_regexp_exclude:
          type: string
          nullable: true
          description: >
            Optional Go regexp; matching titles are not indexed. Empty = no exclude filter.
            Takes priority over title_regexp_include when both match.
        scan_cutoff:
          type: string
          nullable: true
          description: Optional date cutoff; empty = index all videos from source (ignored for kind=single)
        last_error_code:
          type: string
          nullable: true
          description: Derived from latest source_history scan_error (sticky); empty after a later scanned
        last_error_message:
          type: string
          nullable: true
          description: Derived from latest source_history scan_error message
        last_scanned_at:
          type: string
          format: date-time
          nullable: true
          description: Derived from created_at of latest source_history scanned or scan_error
...
    Video:
      type: object
      required: [id, series_id, remote_id, title, status]
      properties:
        id:
          type: integer
          format: int64
        series_id:
          type: integer
          format: int64
        source_id:
          type: integer
          format: int64
          nullable: true
        remote_id:
          type: string
        title:
          type: string
        upload_date:
          type: string
          nullable: true
          description: RFC3339 UTC when known
        source_url:
          type: string
          nullable: true
          description: Video page URL (watch/clip); never the feed channel/playlist URL
        status:
          type: string
          description: wanted, wanted_source_error, wanted_download_error, downloaded, verify_failed, missing, deleted, ignored, streamable
        season:
          type: integer
          nullable: true
        episode:
          type: integer
          nullable: true
        description:
          type: string
        thumbnail_url:
          type: string
          nullable: true
        media_type:
          type: string
          nullable: true
          description: yt-dlp media_type when known (e.g. video, short, livestream, clip); empty/null when missing
        size_bytes:
          type: integer
          format: int64
          nullable: true
          description: On-disk size of the packed video media file (kind=video); null when unknown or not downloaded
        resolution:
          type: string
          nullable: true
          description: >
            Rough resolution bucket from min(width,height): 240p, 360p, 480p, 720p, 1080p, or 4K.
            1440-class short side maps to 1080p. Null when width/height unknown.
          enum: [240p, 360p, 480p, 720p, 1080p, 4K]
...
    Series:
      type: object
      required: [id, title, root_id, quality_profile_id, monitored, added_at]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        root_id:
          type: integer
          format: int64
        quality_profile_id:
          type: integer
          format: int64
        monitored:
          type: boolean
          description: Series-level monitored; never rewrites source/video/domain monitored flags
        auto_ignore_media_types:
          type: array
          items:
            type: string
          description: >
            yt-dlp media_type values to auto-ignore (e.g. short, livestream). Empty = none.
            Missing media_type is never auto-ignored. Applied at index create when type is known,
            at download via --match-filters, and at stream pack_stream from the urls extract
            (indexed, then marked ignored; no cache_beginning).
        added_at:
          type: string
          format: date-time
        root_name:
          type: string
        quality_profile_name:
          type: string
        video_count:
          type: integer
          format: int64
        downloaded_count:
          type: integer
          format: int64
        wanted_count:
          type: integer
          format: int64
        source_count:
          type: integer
          format: int64
        sources:
          type: array
          items:
            $ref: "#/components/schemas/Source"
        videos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
          description: Present on series detail GET only
...